Move Sharees api over to use OCSController
